Added the basics for the About Me section of the portfolio
Added the basics for the About Me section of the portfolio, still needs minor style tweaking and will need to be linked up to a database for optimal use and to reflect the most recent version of this description. However this is good enough for a first release of the page.